+++ Example/Cheats/ViewController.swift
@@ -15,6 +15,9 @@ class ViewController: UIViewController {
     @IBOutlet weak var cheatStateLabel: UILabel!
     @IBOutlet weak var nextActionLabel: UILabel!
 
+    // MARK: State
+    private var gestureRecognizer: CheatCodeGestureRecognizer?
+
     // MARK: Life Cycle
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -35,6 +38,7 @@ class ViewController: UIViewController {
         // Add the gesture recognizer
         let gestureRecognizer = CheatCodeGestureRecognizer(cheatCode: cheatCode, target: self,
                                                            action: #selector(actionPerformed(_:)))
+        self.gestureRecognizer = gestureRecognizer
         view.addGestureRecognizer(gestureRecognizer)
     }
 
@@ -47,14 +51,12 @@ class ViewController: UIViewController {
 // MARK: Cheat Codes
 extension ViewController {
 
-    /// Reset the UI after cheat code correct / incorrect
-    func reset(cheatCode: CheatCode) {
-        let deadline = DispatchTime.now() + 2.0
-        DispatchQueue.main.asyncAfter(deadline: deadline) {
-            cheatCode.reset()
-            self.cheatStateLabel.text = "Cheat sequence reset"
-            self.updateNextActionLabel(cheatCode: cheatCode)
-        }
+    func showAlert(message: String?) {
+        gestureRecognizer?.configureForNextAction()
+        let alertController = UIAlertController(title: message, message: nil, preferredStyle: .alert)
+        let okAction = UIAlertAction(title: "OK", style: .default)
+        alertController.addAction(okAction)
+        present(alertController, animated: true, completion: nil)
     }
 
     /// Update UI to indicate cheat sequence progress
@@ -62,12 +64,14 @@ extension ViewController {
         switch cheatCode.state() {
         case .matched:
             self.cheatStateLabel.text = "Cheat unlocked!"
-            self.reset(cheatCode: cheatCode)
+            showAlert(message: self.cheatStateLabel.text)
         case .matching:
             self.cheatStateLabel.text = "Cheat incomplete"
         case .notMatched:
             self.cheatStateLabel.text = "Cheat incorrect"
-            self.reset(cheatCode: cheatCode)
+            showAlert(message: self.cheatStateLabel.text)
+        case .reset:
+            self.cheatStateLabel.text = "Cheat sequence reset"
         }
     }